additional changes for ansible 2.4 support (fixes probe process)
For a dynamic inventory under ansible 2.3 you could pass hosts=[] when
creating an Inventory instance. In 2.4, this changes. Ansible 2.4 supports
multiple inventory sources, so hosts= is replaced by sources=.

Furthermore, if a list is passed to sources the api treats each item as a
inventory source file to read! To match the previous behavior the sources
attribute needs to be a comma separated list of hosts.