new module added to handle host access interaction
This module creates the widgets for the host access page, where the hosts
entered are checked for passwordless ssh configuration. The UI effectively
presents two tables; a pending table and an access ok table. When the
user first enters the page, *all* hosts are in the pending state. The user
then clicks on 'Check', which tests for passwordless ssh (root only).

If access is OK, the host moves to the "access ok" table. If not, the
user can provide a generic root password for all hosts or specify them
individually and click 'Check' again. If the password provided is correct
copilot will use the root credentials to setup passwordless ssh between
the ansible node and the candidate ceph node.

Once all hosts appear in the "Access Ok" table, the workflow continues.